requirements.txt
requests==2.28.1
ruamel.yaml==0.17.21
pyTelegramBotAPI==4.6.0